+++ src/document.cpp
@@ -18,11 +18,11 @@
 #include <vector>
 #include <wx/tokenzr.h>
 
-int document::next_section_index(long position) const noexcept {
+int document::next_section_index(int position) const noexcept {
 	return buffer.next_marker_index(position, marker_type::section_break);
 }
 
-int document::previous_section_index(size_t position) const noexcept {
+int document::previous_section_index(int position) const noexcept {
 	return buffer.previous_marker_index(position, marker_type::section_break);
 }
 
@@ -34,11 +34,11 @@ size_t document::offset_for_section(int section_index) const noexcept {
 	return buffer.marker_position(section_index);
 }
 
-int document::next_page_index(long position) const noexcept {
+int document::next_page_index(int position) const noexcept {
 	return buffer.next_marker_index(position, marker_type::page_break);
 }
 
-int document::previous_page_index(size_t position) const noexcept {
+int document::previous_page_index(int position) const noexcept {
 	return buffer.previous_marker_index(position, marker_type::page_break);
 }
 
@@ -73,11 +73,11 @@ int document::find_closest_toc_offset(size_t position) const noexcept {
 	return best_offset;
 }
 
-int document::next_heading_index(long position, int level) const noexcept {
+int document::next_heading_index(int position, int level) const noexcept {
 	return buffer.next_heading_marker_index(position, level);
 }
 
-int document::previous_heading_index(size_t position, int level) const noexcept {
+int document::previous_heading_index(int position, int level) const noexcept {
 	return buffer.previous_heading_marker_index(position, level);
 }
